Wave 7 matching: low-band structural fixes, many big jumps
Focused on the 55-90% band (structural/correctness, not register perms).
Highlights: RomCurve_countRandomPoints 66->99.6 (inline noUnblockedLinks),
expgfx_renderSourcePools 63->100 (direct subscript vs helper), GameUI_release
70->90.6 (extern-array base+offset dodges lwzx isel gap), mapBlockRender_
setupShaderTextures 74->90, trickyFn_80142eb0/801434b0 76/82->98.8/98.6
(switch dispatch restructure), player value-arg VN decoupling, RomCurve binary-
search alignment (lifted 16 inline sites). Recipes #11/#13/#28/#33/#39/#46/#59/
#67b/#74/#91/#112/#114. Per-file 3-way merge onto origin (curves/tex convergent).

+++ src/main/voxmaps.c
@@ -32,7 +32,7 @@ int* voxmaps_getRouteNode(u8* header, int* nodeBase, u8* bitmap, int d, int e, i
         }
         cur++;
     }
-    byte = *end & ((u32)0xff >> (8 - (d & 7)));
+    byte = *end & (u8)((u32)0xff >> (8 - (d & 7)));
     while (byte != 0)
     {
         byte &= byte - 1;
